refactor: transforma Icon em componente global
