+++ src/components/PlanetIcon/index.js
@@ -1,6 +1,6 @@
-import Planeticon from "../../../../assets/Frame 5.png"
-import Polygon1 from "../../../../assets/Frame 3.png"
-import Polygon2 from "../../../../assets/Frame 4.png"
+import Planeticon from "../../assets/Frame 5.png"
+import Polygon1 from "../../assets/Frame 3.png"
+import Polygon2 from "../../assets/Frame 4.png"
 
 import { Container, Planet, PurplePolygon, WhitePolygon } from "./style";
