+++ src/components/PlanetIcon/style.js
